% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{c_ngrams}
\alias{c_ngrams}
\title{Concatenate ngrams in a given character vector}
\usage{
c_ngrams(data, n_max, text_col = NULL, ...)
}
\arguments{
\item{data}{Data frame including text column for which ngrams should be concatenated}

\item{n_max}{Maximum size of ngrams to concatenate}

\item{text_col}{Name of the text column to search for ngrams}

\item{...}{Arguments to be passed to the nested \code{get_ngrams()} function}
}
\value{
Data frame with text column containing concatenated ngrams
}
\description{
This function identifies ngrams in a text and replaces them with their
concatenated and lemmatised version.
}
