% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_basic}
\alias{clean_basic}
\title{Clean body of text}
\usage{
clean_basic(text)
}
\arguments{
\item{text}{A vector string to be converted}
}
\value{
A cleaned vector string
}
\description{
Function removes white space, punctuation and  converts the strings to lowercase
}
