% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{convert_to_dtm}
\alias{convert_to_dtm}
\title{Convert data frame with word frequencies into document term matrix}
\usage{
convert_to_dtm(data, title_col = NULL, word_col = NULL, min_term_freq = 1)
}
\arguments{
\item{data}{Data frame containing colomns with titles, words and word counts}

\item{title_col}{Name of the title column of the input data frame}

\item{word_col}{Name of the word column of the input data frame}
}
\value{
A document term matrix
}
\description{
Convert data frame with word frequencies into document term matrix
}
