% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{count_docs_per_topic}
\alias{count_docs_per_topic}
\title{Count documents per main topic}
\usage{
count_docs_per_topic(lda)
}
\arguments{
\item{lda}{LDA model}
}
\value{
Data frame with topics and count of documents per main topic
}
\description{
Count documents per main topic
}
