% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emu_metadata}
\alias{emu_metadata}
\title{EMU theses  metadata}
\format{
\subsection{\code{emu_metadata}}{

A data frame with 68 rows and 10 columns:
\describe{
\item{ID}{ID of the thesis}
\item{grad_year}{The year when the thesis was published.}
\item{grad_sem}{The semester when the student graduated.}
\item{full_title}{Full title of the thesis.}
\item{title}{Title of the thesis.}
\item{subtitle}{Subtitle of the thesis.}
\item{loc}{The location that the thesis is focused on.}
\item{lat}{Latitude of the thesis location geocoded from \code{location}.}
\item{long}{Longitude of the thesis location geocoded from \code{location}.}
\item{abstract}{Abstract of the thesis.}
}
}
\source{
\url{https://repository.tudelft.nl/islandora/search/?collection=education}
}
\usage{
emu_metadata
}
\description{
A data frame with metadata of theses from the EMU program.
The data frame can be joined with the \code{emu_raw} data frame via ID column.
}
\keyword{datasets}
