% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emu_raw}
\alias{emu_raw}
\title{EMU theses raw text data}
\format{
\subsection{\code{emu_raw}}{

A data frame with 83 rows and 2 columns:
\describe{
\item{ID}{ID of the thesis}
\item{text_raw}{Raw text of the thesis.
The text has been anonymyzed and Latin special characters have been replaced with 26-letter alphabet counterparts}
}
}
\source{
\url{https://repository.tudelft.nl/islandora/search/?collection=education}
}
\usage{
emu_raw
}
\description{
A data frame with text of theses from the EMU program.
The data frame can be joined with the \code{emu_metadata} data frame via ID column.
}
\keyword{datasets}
