% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{find_meta_stopwords}
\alias{find_meta_stopwords}
\title{Find custom stop-phrases based on metadata file}
\usage{
find_meta_stopwords(
  metadata,
  stop_cols = list(author1 = c("first_name", "last_name"), author2 = c("first_name_2",
    "last_name_2"), title = "title", subtitle = "subtitle", full_title = "full_title"),
  convert_to_regex = TRUE
)
}
\arguments{
\item{metadata}{A data frame with metadata information about the text}

\item{stop_cols}{List of names of columns that should be used for extraction of the custom stop words, the vectors need to be named as in the example}

\item{convert_to_regex}{A logical value determining if the final outcome should be a regex input. Default is TRUE}
}
\value{
A list with strings to be removed for each of the texts
}
\description{
Function scans the dataset with metadata for fields that can be removed from the body fo text before the text analysis
}
