% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{fit_lda}
\alias{fit_lda}
\title{Fit LDA model}
\usage{
fit_lda(dtm, k, ...)
}
\arguments{
\item{dtm}{Document term matrix}

\item{k}{Number of topics}

\item{...}{Additional arguments to be passed to the LDA function}
}
\value{
A list with the beta and theta matrices
}
\description{
Fit LDA model
}
