% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode_thesis_locations}
\alias{geocode_thesis_locations}
\title{Geocode thesis locations}
\usage{
geocode_thesis_locations(data, location = NULL)
}
\arguments{
\item{data}{Data frame with thesis data}
}
\value{
Data frame with latitude and longitude columns
}
\description{
The function creates two new columns, latitude and longitude,
based on a column containing the address of the thesis.
}
