% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_doc_counts_per_topic}
\alias{get_doc_counts_per_topic}
\title{Count documents by main topic}
\usage{
get_doc_counts_per_topic(theta)
}
\arguments{
\item{theta}{Matrix with each row representing the probability
distribution of a document over topics}
}
\value{
Data frame with topics and count of documents per each topic
}
\description{
For each topic, count how many documents have it as the main topic.
}
