% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_max_topic_per_doc}
\alias{get_max_topic_per_doc}
\title{Get dominant topic per document}
\usage{
get_max_topic_per_doc(theta)
}
\arguments{
\item{theta}{Matrix with each row representing the probability distribution
of a document over topics.}
}
\value{
Tibble with the dominant topic per document
}
\description{
Get dominant topic per document
}
