% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_number_of_topics}
\alias{get_number_of_topics}
\title{Get number of topics for LDA model empirically}
\usage{
get_number_of_topics(
  dtm,
  topics = seq(from = 2, to = 30, by = 1),
  metrics = c("CaoJuan2009", "Deveaud2014"),
  method = "Gibbs",
  control = list(seed = 2023),
  verbose = TRUE
)
}
\arguments{
\item{dtm}{Document term matrix}

\item{topics}{Range of number of topics}

\item{metrics}{Metrics used to evaluate the number of topics}

\item{method}{Method used for fitting the LDA model}

\item{control}{Setting tuning parameters}

\item{verbose}{Show progress in the console}
}
\value{
Plot and possibly a K value
}
\description{
Get number of topics for LDA model empirically
}
