% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_tf_idf}
\alias{get_tf_idf}
\title{Get \code{tf-idf} statistic}
\usage{
get_tf_idf(data, id_col = NULL, word_col = NULL)
}
\arguments{
\item{data}{Data frame containing colomns with IDs, words and word counts}

\item{id_col}{Name of the ID column of the input data frame}

\item{word_col}{Name of the word column of the input data frame}
}
\value{
A data frame including the \code{tf-idf} statistic for each word
}
\description{
Get \code{tf-idf} statistic
}
