% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_top_words_per_document}
\alias{get_top_words_per_document}
\title{Extract top n words with the highest count per document}
\usage{
get_top_words_per_document(
  data,
  top_n,
  id_col = "ID",
  word_col = "word",
  min_count = 1
)
}
\arguments{
\item{data}{Data frame with theses, containing a column with words.}

\item{top_n}{The number of top words to be returned by the function.}

\item{id_col}{The name of the column from the input data frame containing the ID of the document.}

\item{word_col}{Name of the column from the input data frame containing all words.}

\item{min_count}{Minimum count to be considered. Words below this value will not be included in the returned data frame.}
}
\value{
A data frame with the top n words with a count above the given minimum.
}
\description{
Given a data frame with a column storing the words found in a document
identified with a ID column and a positive number top_n, this function
returns the top n words with the highest count in each document.
}
