% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_top_words_per_topic}
\alias{get_top_words_per_topic}
\title{Extract top n terms with the highest count across all topics}
\usage{
get_top_words_per_topic(data, top_n, topic_col = "topic", beta_col = "beta")
}
\arguments{
\item{data}{Data frame with topics, terms and beta statistic.}

\item{top_n}{The number of top words to be returned by the function.}

\item{topic_col}{Name of the column containing topics.}

\item{beta_col}{Name of the column containing the beta statistic for all terms.}
}
\value{
A data frame with the top n terms across all topics.
}
\description{
Extract top n terms with the highest count across all topics
}
