% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_topic_proportions}
\alias{get_topic_proportions}
\title{Get topic proportions in an LDA model}
\usage{
get_topic_proportions(theta, digits = 2)
}
\arguments{
\item{theta}{Matrix with each row representing the probability
distribution of a document over topics}

\item{digits}{Number of digits to round the results}
}
\value{
Data frame with topic ids and their proportions in the corpus
}
\description{
Calculate the mean probability of each topic over all documents.
}
