% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_topics}
\alias{get_topics}
\title{Get topics for a given data frame of documents and words}
\usage{
get_topics(
  data,
  id_col = "ID",
  word_col = "word",
  k = 2,
  min_term_freq = 1,
  seed = 2023,
  ...
)
}
\arguments{
\item{data}{Data frame with theses, containing a column with words.}

\item{id_col}{The name of the column from the input data frame containing the ID of the document.}

\item{word_col}{Name of the column from the input data frame containing all words.}

\item{k}{Number of desired clusters.}

\item{seed}{Seed to control pseudorandomisation.}

\item{matrix}{Statistics to be computed. Can be "beta" or "gamma"}
}
\value{
An LDA model object.
}
\description{
Get topics for a given data frame of documents and words
}
