% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_word_counts}
\alias{get_word_counts}
\title{Get word counts for corpus}
\usage{
get_word_counts(data, text_col = NULL)
}
\arguments{
\item{data}{A data frame comprising a corpus of texts}

\item{text_col}{Name of the column in which the text is stored}
}
\value{
A numeric vector of word counts, one for each document in the corpus
}
\description{
Get word counts for corpus
}
