% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{head_text}
\alias{head_text}
\title{Show the beginning of each of the element of a string vector}
\usage{
head_text(text_vector, head_n = 10, sub_start = 1, sub_end = 500)
}
\arguments{
\item{text_vector}{A string vector}

\item{head_n}{Number of elements of the vector. Default is 10}

\item{sub_start}{starting position within the string. Default is 1.}

\item{sub_end}{final position within the string. Default is 500.}
}
\value{
A initial characters of  first elements of the string vector
}
\description{
Show the beginning of each of the element of a string vector
}
