% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{name_topics}
\alias{name_topics}
\title{Add pseudo-names to topics in an LDA model}
\usage{
name_topics(beta, n_words = 5)
}
\arguments{
\item{beta}{Prior topic distribution over words}

\item{n_words}{Number of words to include in the pseudo-names}
}
\value{
Character vector of topic pseudo-names
}
\description{
Pseudo-names are obtained by pasting together the top n terms with
the highest probability within a topic.
}
