% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{nearest_neighbors}
\alias{nearest_neighbors}
\title{Get nearest neighbours for a given token}
\usage{
nearest_neighbors(df, token)
}
\arguments{
\item{df}{Data frame with word vectors}

\item{token}{Word for which the nearest neighbours are calculated}
}
\value{
Data frame with nearest neighbours
}
\description{
Get nearest neighbours for a given token
}
