% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{normalise_words}
\alias{normalise_words}
\title{Check for special Latin letters}
\usage{
normalise_words(words)
}
\arguments{
\item{words}{A vector string to be converted. each element will be treated as a separate word}
}
\value{
A vector string with regular letter counterparts added at the end of the string
}
\description{
Function checks if the vector of words includes special Latin letters;
if so, add a word with a regular Latin counterpart to the vector
}
