% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{short_words}
\alias{short_words}
\title{Produces the table with the most popular short words in the body of text}
\usage{
short_words(
  text_vector,
  l_word = 4,
  l_list = 50,
  scope = "word",
  print_kable = TRUE
)
}
\arguments{
\item{text_vector}{A string vector}

\item{l_word}{length of words to be detected}

\item{l_list}{number of top words to be retrieved}

\item{scope}{either 'word' or 'corpus'. should be search be performed on individual words or a corpus?}

\item{print_kable}{logical value. If TRUE (default), a kable with results is printed out}
}
\value{
A table with the short words and its prevalance in the body of text
}
\description{
Function creates a vector of stopwords associated with master theses
}
