% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{slide_windows}
\alias{slide_windows}
\title{Sliding window}
\usage{
slide_windows(tbl, window_size)
}
\arguments{
\item{tbl}{A vector of tokens.}

\item{window_size}{A small window size of 3-4 shows how a word is used, while a larger window size of, say, 10 gives more information about the domain topic.}
}
\value{
Data frame with skipgram probabilities
}
\description{
Sliding window function used to calculate skipgram probabilities,
that is, how often each word occurs near each other word.
}
