% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{subject_stopwords}
\alias{subject_stopwords}
\title{Adjustable list of stopwords associated with urbanism}
\usage{
subject_stopwords(add_stopwords = NULL)
}
\arguments{
\item{add_stopwords}{A vector of words to be added to the urbanism stopwords}
}
\value{
A vector of urbanism-related stopwords
}
\description{
Function creates a vector of stopwords associated with field of urbanism
}
