% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tbl_var_overview}
\alias{tbl_var_overview}
\title{Create table with variable overview}
\usage{
tbl_var_overview(data, html = FALSE)
}
\arguments{
\item{data}{Data frame}

\item{html}{Logical, should the output be in HTML format?}
}
\value{
A tibble with columns \code{Column}, \code{Type}, \code{Completeness}
}
\description{
Create table with variable overview
}
