% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{thesis_stopwords}
\alias{thesis_stopwords}
\title{Adjustable list of stopwords associated with master theses}
\usage{
thesis_stopwords(add_stopwords = NULL)
}
\arguments{
\item{add_stopwords}{A vector of words to be added to the master theses-related stopwords}
}
\value{
A vector of thesis-related stopwords
}
\description{
Function creates a vector of stopwords associated with master theses
}
