% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{tokenize}
\alias{tokenize}
\title{Tokenise a given corpus}
\usage{
tokenize(
  data,
  id_col = "ID",
  text_col = "text_clean",
  min_length = 3,
  lemma = TRUE
)
}
\arguments{
\item{data}{A data frame containing a text column}

\item{id_col}{Unique document id}

\item{text_col}{Name of the column containing the text to be tokenized}

\item{min_length}{Minimum length of words to be kept}

\item{lemma}{Lemmatise words}
}
\value{
Data frame with \code{ID} and \code{text_clean} columns
}
\description{
Tokenise a given corpus
}
