% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{top_n_word_count_per_corpus}
\alias{top_n_word_count_per_corpus}
\title{Find top n words with the highest count across all documents}
\usage{
top_n_word_count_per_corpus(data, top_n, word_col = NULL)
}
\arguments{
\item{data}{Data frame with theses, containing a column with words.}

\item{top_n}{The number of top words to be returned by the function.}

\item{word_col}{Name of the column from the input data frame containing all words.}
}
\value{
A data frame with the top n words across all documents.
}
\description{
Given a data frame with a column storing the words found in a document
and a positive number top_n, this function returns the top n words with the
highest count across all documents.
}
