% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{urbanism_stopwords}
\alias{urbanism_stopwords}
\title{Adjustable list of stopwords associated with urbanism}
\usage{
urbanism_stopwords(add_stopwords = NULL, convert_to_regex = TRUE)
}
\arguments{
\item{add_stopwords}{A vector of words to be added to the urbanism stopwords}
}
\value{
A vector of urbanism-related stopwords
}
\description{
Function creates a vector of stopwords associated with field of urbanism
}
