% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{vis_dom_topic_distr}
\alias{vis_dom_topic_distr}
\title{Visualise topic distributions over the years}
\usage{
vis_dom_topic_distr(data, max_topic_col = NULL, year_col = NULL)
}
\arguments{
\item{data}{Data frame with theses, containing columns with topics and years.}

\item{max_topic_col}{Name of the column containing the main topic.}

\item{year_col}{Name of the column containing the year.}
}
\value{
\code{ggplot} boxplot of topic distributions over the years.
}
\description{
Visualise topic distributions over the years
}
