% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{vis_top_words_per_corpus}
\alias{vis_top_words_per_corpus}
\title{Visualise top words per corpus}
\usage{
vis_top_words_per_corpus(lda, words, top_n = 20)
}
\arguments{
\item{...}{Arguments to be passed to \code{get_top_words_per_corpus()}.}
}
\value{
\code{ggplot} bar chart of top n words colored by topic.
}
\description{
Visualise top words per corpus
}
