% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{vis_top_words_per_topic}
\alias{vis_top_words_per_topic}
\title{Visualise top words per topic}
\usage{
vis_top_words_per_topic(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{get_top_words_per_topic()}.}
}
\value{
\code{ggplot} bar chart of top n words faceted by topic.
}
\description{
Visualise top words per topic
}
