% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{visualise_first_n_components}
\alias{visualise_first_n_components}
\title{Visualise first n components}
\usage{
visualise_first_n_components(word_vectors, dim = 24, top_n = 12)
}
\arguments{
\item{word_vectors}{Word vectors}

\item{dim}{Number of dimensions to visualise}

\item{top_n}{Number of top terms to show for each dimension}
}
\value{
ggplot object
}
\description{
Visualise first n components
}
