% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{visualise_lda}
\alias{visualise_lda}
\title{Visualise LDA model}
\usage{
visualise_lda(phi, theta, dtm)
}
\arguments{
\item{phi}{A matrix with each row representing the probability
distribution of a topic over terms}

\item{theta}{A matrix with each row representing the probability
distribution of a documents over topics}

\item{dtm}{Document-term matrix used as input to the LDA model}
}
\value{
Visualise LDA in a browser.
}
\description{
Visualise LDA model
}
