% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{visualize_thesis_locations}
\alias{visualize_thesis_locations}
\title{Visualise thesis locations}
\usage{
visualize_thesis_locations(data)
}
\arguments{
\item{data}{Data frame with thesis data.}
}
\value{
A plot with thesis locations
}
\description{
Show thesis locations on the map of the world.
}
