% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{word_vec_combos}
\alias{word_vec_combos}
\title{Create a combination of words}
\usage{
word_vec_combos(string1, string2, convert_to_regex = TRUE)
}
\arguments{
\item{string1}{A vector string to be included in the combination}

\item{string2}{A vector string to be included in the combination}
}
\value{
A vector string consisting of combinations of words coming from input vectors
}
\description{
Function creates a vector that includes combination of words
coming from two vector of strings,
making sure at least one word from each input strings is included in the combination
}
