% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_pdf_text}
\alias{convert_pdf_text}
\title{Convert pdf to text}
\usage{
convert_pdf_text(pdf_filenames, by_page = TRUE)
}
\arguments{
\item{pdf_filenames}{A vector of paths to input (pdf) files}
}
\value{
A vector of body of text; one string per pdf file
}
\description{
Function returns a list of files from the directory chosen by the user and to be converted.
Checks if the provided file path exists and whether it holds the adequate file formats (pdf).
}
