% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_ngrams}
\alias{get_ngrams}
\title{Get ngrams from a corpus}
\usage{
get_ngrams(
  data,
  n,
  title_col = NULL,
  text_col = NULL,
  rm_n_short_words = 2,
  stem = FALSE,
  lemma = FALSE
)
}
\arguments{
\item{data}{Data frame of documents containing a text column}

\item{n}{The number of words to include in the token (e.g., n = 2 for bigram)}

\item{title_col}{The name of the title column}

\item{text_col}{The name of the column where the text to be tokenised is located}

\item{stem}{If \code{TRUE} create additional columns with stemmed version of tokens}

\item{lemma}{If \code{TRUE} create additional columns with lemmatised version of tokens}
}
\value{
A data frame with an "ngram" column and columns for its constituent words
}
\description{
Get ngrams from a corpus
}
