% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_pmi}
\alias{get_pmi}
\title{Pointwise mutual information of pairs of items}
\usage{
get_pmi(
  data,
  id_col = NULL,
  text_col = NULL,
  parallel = TRUE,
  window_size = 4L
)
}
\arguments{
\item{data}{Data frame of theses}

\item{id_col}{Name of the ID column}

\item{text_col}{Name of the text column to be tokenised}

\item{parallel}{If \code{parallel = TRUE}, compute in parallel}
}
\value{
Data frame with two items and their co-occurrence
}
\description{
Pointwise mutual information of pairs of items
}
