% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_topic_proportions}
\alias{get_topic_proportions}
\title{Get topic proportions in an LDA model}
\usage{
get_topic_proportions(lda, digits = 2)
}
\arguments{
\item{lda}{LDA model object}
}
\value{
Data frame with topics and their proportions
}
\description{
Get topic proportions in an LDA model
}
