% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_topics}
\alias{get_topics}
\title{Get topics for a given data frame of documents and words}
\usage{
get_topics(data, k = 2, seed = 2023, title_col = "title", word_col = "word")
}
\arguments{
\item{data}{Data frame with theses, containing a column with words.}

\item{k}{Number of desired clusters.}

\item{seed}{Seed to control pseudorandomisation.}

\item{title_col}{The name of the column from the input data frame containing the title of the document.}

\item{word_col}{Name of the column from the input data frame containing all words.}

\item{matrix}{Statistics to be computed. Can be "beta" or "gamma"}
}
\value{
An LDA model object.
}
\description{
Get topics for a given data frame of documents and words
}
